[clang][contracts] contract options, tokens and diagnostics
The vocabulary every other contracts commit is written against: the language
options, the token kinds, the contract and semantic enumerations, and the
diagnostic groups.

Nothing here does anything on its own, which is the point -- it is the
header layer that lets the AST, parser, Sema and CodeGen commits that follow
refer to the same types.

One thing in here is worth a reviewer's attention: the
`AllContractSemanticsMask*` constants name P3100's `assume` and P4298's
noexcept-terminating semantics even though neither paper has landed at this
point in the series.  They are masks over the base enumeration, and the
comment says explicitly that the gates are applied elsewhere.

+++ clang/include/clang/Basic/IdentifierTable.h
@@ -58,7 +58,7 @@ enum KeywordStatus {
 
 /// Translates flags as specified in TokenKinds.def into keyword status
 /// in the given language standard.
-KeywordStatus getKeywordStatus(const LangOptions &LangOpts, unsigned Flags);
+KeywordStatus getKeywordStatus(const LangOptions &LangOpts, uint64_t Flags);
 
 enum class ReservedIdentifierStatus {
   NotReserved = 0,
